Move custom reusable structs into domain package
