+++ domain/port.go
@@ -1,4 +1,4 @@
-package firewall
+package domain
 
 import (
 	"errors"
